Reported VSO-1659408 '/clr alignas emits error C2711 "this function cannot be compiled as managed" instead of falling back to native codegen'.
