+++ tests/std/tests/P0896R4_ranges_alg_uninitialized_copy/test.cpp
@@ -77,16 +77,6 @@ struct int_wrapper {
     auto operator<=>(const int_wrapper&) const = default;
 };
 
-template <class T, size_t N>
-struct holder {
-    STATIC_ASSERT(N < ~size_t{0} / sizeof(T));
-    alignas(T) unsigned char space[N * sizeof(T)];
-
-    auto as_span() {
-        return span<T, N>{reinterpret_cast<T*>(space + 0), N};
-    }
-};
-
 struct instantiator {
     static constexpr int expected_output[]      = {13, 55, 12345};
     static constexpr int expected_output_long[] = {13, 55, 12345, -1};
