+++ tests/std/tests/P1135R6_atomic_flag_test/test.cpp
@@ -82,6 +82,8 @@ void test_flag_type() {
 }
 
 int main() {
+#ifndef _M_CEE // TRANSITION, VSO-1659408
     test_flag_type<std::atomic_flag>();
     test_flag_type<volatile std::atomic_flag>();
+#endif // _M_CEE
 }
